add plots test utils
