+++ extension/src/test/suite/experiments/model/tree.test.ts
@@ -1,19 +1,13 @@
 import { afterEach, beforeEach, describe, it, suite } from 'mocha'
 import { expect } from 'chai'
-import { spy, stub, restore, SinonStub } from 'sinon'
+import { stub, restore, SinonStub } from 'sinon'
 import { commands, QuickPickItem, QuickPickOptions, window } from 'vscode'
 import { Disposable } from '../../../../extension'
-import { WorkspaceExperiments } from '../../../../experiments/workspace'
 import { Status } from '../../../../experiments/model'
 import { dvcDemoPath } from '../../util'
 import { RegisteredCommands } from '../../../../commands/external'
-import { buildExperiments } from '../util'
-import { WorkspacePlots } from '../../../../plots/workspace'
-import { BaseWebview } from '../../../../webview'
-import { CliReader } from '../../../../cli/reader'
-import { Plots } from '../../../../plots'
+import { buildPlots, getExpectedData } from '../../plots/util'
 import livePlotsFixture from '../../../fixtures/expShow/livePlots'
-import { LivePlotsData } from '../../../../plots/webview/contract'
 
 suite('Experiments Tree Test Suite', () => {
   const disposable = Disposable.fn()
@@ -27,6 +21,14 @@ suite('Experiments Tree Test Suite', () => {
   })
 
   describe('ExperimentsTree', () => {
+    const { colors } = livePlotsFixture
+    const ids = [
+      '4fb124aebddb2adf1545030907687fa9a4c80e70',
+      '42b8736b08170529903cd203a1f40382a4b4a8cd',
+      '1ba7bcd6ce6154e72e18b155475663ecbbd1f49d'
+    ]
+    const { domain, range } = colors
+
     it('should appear in the UI', async () => {
       await expect(
         commands.executeCommand('dvc.views.experimentsTree.focus')
@@ -34,48 +36,11 @@ suite('Experiments Tree Test Suite', () => {
     })
 
     it('should be able to toggle whether an experiment is shown in the plots webview with dvc.views.experimentsTree.toggleStatus', async () => {
-      const { experiments, internalCommands, resourceLocator } =
-        buildExperiments(disposable)
+      const { plots, messageSpy } = await buildPlots(disposable)
 
-      await experiments.isReady()
-
-      const messageSpy = spy(BaseWebview.prototype, 'show')
-
-      stub(CliReader.prototype, 'plotsShow').resolves({})
-
-      const plots = disposable.track(
-        new Plots(dvcDemoPath, internalCommands, resourceLocator.scatterGraph)
-      )
-      plots.setExperiments(experiments)
-      await plots.isReady()
-
-      stub(WorkspaceExperiments.prototype, 'getRepository').returns(experiments)
-      stub(WorkspacePlots.prototype, 'getRepository').returns(plots)
-
-      const { plots: plotsData, colors } = livePlotsFixture
-      const ids = [
-        '4fb124aebddb2adf1545030907687fa9a4c80e70',
-        '42b8736b08170529903cd203a1f40382a4b4a8cd',
-        '1ba7bcd6ce6154e72e18b155475663ecbbd1f49d'
-      ]
-
-      const { domain, range } = colors
       const expectedDomain = [...domain]
       const expectedRange = [...range]
       const expectedIds = [...ids]
-
-      const getExpectedLivePlots = (): LivePlotsData => ({
-        colors: {
-          domain: expectedDomain,
-          range: expectedRange
-        },
-        plots: plotsData.map(plot => ({
-          title: plot.title,
-          values: plot.values.filter(values =>
-            expectedDomain.includes(values.group)
-          )
-        }))
-      })
 
       await plots.showWebview()
 
@@ -83,10 +48,7 @@ suite('Experiments Tree Test Suite', () => {
         expect(
           messageSpy,
           'a message is sent with colors for the currently selected experiments'
-        ).to.be.calledWith({
-          live: getExpectedLivePlots(),
-          static: undefined
-        })
+        ).to.be.calledWith(getExpectedData(expectedDomain, expectedRange))
         messageSpy.resetHistory()
 
         const id = expectedIds.pop()
@@ -128,41 +90,16 @@ suite('Experiments Tree Test Suite', () => {
         Status.selected
       )
 
-      expect(messageSpy, 'we no longer send undefined').to.be.calledWith({
-        live: getExpectedLivePlots(),
-        static: undefined
-      })
+      expect(messageSpy, 'we no longer send undefined').to.be.calledWith(
+        getExpectedData(expectedDomain, expectedRange)
+      )
     }).timeout(6000)
 
     it('should be able to select / de-select experiments using dvc.views.experimentsTree.selectExperiments', async () => {
-      const { experiments, internalCommands, resourceLocator } =
-        buildExperiments(disposable)
-
-      await experiments.isReady()
-
-      const messageSpy = spy(BaseWebview.prototype, 'show')
-
-      stub(CliReader.prototype, 'plotsShow').resolves({})
-
-      const plots = disposable.track(
-        new Plots(dvcDemoPath, internalCommands, resourceLocator.scatterGraph)
-      )
-      plots.setExperiments(experiments)
-      await plots.isReady()
-
-      stub(WorkspaceExperiments.prototype, 'getDvcRoots').returns([dvcDemoPath])
-      stub(WorkspaceExperiments.prototype, 'getRepository').returns(experiments)
-      stub(WorkspacePlots.prototype, 'getRepository').returns(plots)
-
-      const { plots: plotsData, colors } = livePlotsFixture
-      const ids = [
-        '4fb124aebddb2adf1545030907687fa9a4c80e70',
-        '42b8736b08170529903cd203a1f40382a4b4a8cd',
-        '1ba7bcd6ce6154e72e18b155475663ecbbd1f49d'
-      ]
+      const { plots, messageSpy } = await buildPlots(disposable)
 
       const selectedId = ids[0]
-      const { domain, range } = colors
+
       const selectedDisplayName = domain[0]
       const selectedColor = range[0]
       const selectedItem = {
@@ -209,21 +146,9 @@ suite('Experiments Tree Test Suite', () => {
       expect(
         messageSpy,
         'a message is sent with colors for the currently selected experiments'
-      ).to.be.calledWith({
-        live: {
-          colors: {
-            domain: [selectedDisplayName],
-            range: [selectedColor]
-          },
-          plots: plotsData.map(plot => ({
-            title: plot.title,
-            values: plot.values.filter(
-              values => values.group === selectedDisplayName
-            )
-          }))
-        },
-        static: undefined
-      })
+      ).to.be.calledWith(
+        getExpectedData([selectedDisplayName], [selectedColor])
+      )
     }).timeout(6000)
   })
 })
